--- src/main/java/com/das/das_backend/domain/user/presentation/UserController2.java
@@ -1,27 +1,0 @@
-package com.das.das_backend.domain.user.presentation;
-
-import com.das.das_backend.domain.user.presentation.dto.request.UserSignUpRequest;
-import com.das.das_backend.domain.user.service.UserSignUpService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
-import javax.validation.Valid;
-
-@RequiredArgsConstructor
-@RequestMapping("/user")
-@RestController
-public class UserController2 {
-
-    private final UserSignUpService userSignUpService;
-
-    @ResponseStatus(HttpStatus.CREATED)
-    @PostMapping("/signup")
-    public void signUp(@RequestBody @Valid UserSignUpRequest request) {
-        userSignUpService.execute(request);
-    }
-
-
-
-
-
-}
